c file

+++ 0x13-more_singly_linked_lists/101-print_listint_safe.c
@@ -8,7 +8,7 @@
 size_t print_listint_safe(const listint_t *head)
 {
 	int num = 0;
-	long int diff;
+	int diff;
 
 	while (head)
 	{
